refactor(ui): move usePrefersReducedMotion to lib/hooks/
