+++ components/HeroSection/HeroSection.hooks.ts
@@ -18,21 +18,3 @@ export const useWordCycler = (words: string[], interval: number, disabled: boole
   // eslint-disable-next-line security/detect-object-injection
   return wordsRef.current[index] ?? wordsRef.current[0] ?? '';
 };
-
-export const usePrefersReducedMotion = (enabled: boolean) => {
-  const [prefersReducedMotion, setPrefersReducedMotion] = useState(false);
-
-  useEffect(() => {
-    if (!enabled || !window.matchMedia) return;
-
-    const mediaQuery = window.matchMedia('(prefers-reduced-motion: reduce)');
-    setPrefersReducedMotion(mediaQuery.matches);
-
-    const handleChange = () => setPrefersReducedMotion(mediaQuery.matches);
-    mediaQuery.addEventListener('change', handleChange);
-
-    return () => mediaQuery.removeEventListener('change', handleChange);
-  }, [enabled]);
-
-  return enabled && prefersReducedMotion;
-};
